feat : 도서정보 임시데이터 mocks로 이동
#145
