+++ src/pages/admin/LibraryManage/LibraryManage.tsx
@@ -4,60 +4,11 @@ import PageTitle from '@components/Typography/PageTitle';
 import StandardTable from '@components/Table/StandardTable';
 import OutlinedButton from '@components/Button/OutlinedButton';
 import { ChildComponent } from '@components/Table/StandardTable.interface';
+import { columns, rows } from '@mocks/LibraryManageApi';
 
 import SearchBook from './SearchBook';
 
-interface BookListInfoColumn {
-  key: keyof BookListInfo;
-  headerName: string;
-}
-
 const LibraryManage = () => {
-  const columns: BookListInfoColumn[] = [
-    {
-      key: 'no',
-      headerName: '번호',
-    },
-    { key: 'title', headerName: '도서명' },
-    { key: 'author', headerName: '저자' },
-    { key: 'total', headerName: '대출현황' },
-    { key: 'information', headerName: '대출정보' },
-    {
-      key: 'enable',
-      headerName: '대출상태',
-    },
-  ];
-
-  const rows: BookListInfo[] = [
-    {
-      id: 1,
-      no: 1,
-      title: '해킹책1',
-      author: '장서윤',
-      total: '3/3',
-      information: '장서윤, 김은지, 송세연',
-      enable: false,
-    },
-    {
-      id: 2,
-      no: 2,
-      title: '보안책2',
-      author: '아주아주 이름 긴 김은지',
-      total: '1/10',
-      information: '누가 빌려갔냐!',
-      enable: false,
-    },
-    {
-      id: 3,
-      no: 3,
-      title: '아주아주 긴 보안책',
-      author: '장서윤',
-      total: '1/1',
-      information: '접니다',
-      enable: true,
-    },
-  ];
-
   const childComponent = ({ key, value, rowData }: ChildComponent<BookListInfo>) => {
     switch (key) {
       case 'enable':
